Replace PYTHONPATH management ...
with a solution based on pipenv and setuptools.
Setuptools is used to define a package from this
package.  This package is then installed into
the virtual environment via pipenv.  A
convenience script run is provided to shorten
the commands to execute the exercises.  The
PYTHONPATH management scripts have been removed.